+++ src/el/ctk_utils/types.py
@@ -23,6 +23,9 @@ type AnchorType = typing.Literal["center", "n", "s", "e", "w"]
 type JustifyType = typing.Literal["left", "center", "right"]
 type CompoundType = typing.Literal["right", "left", "top", "bottom"]
 type StateType = typing.Literal["normal", "disabled", "active"]
+type OrientationType = typing.Literal["vertical", "horizontal"]
+type ViewCommand = typing.Literal["moveto", "scroll"]
+
 
 class GridRowColConfigArgs(typing.TypedDict, total=False):
     minsize: int