+++ src/el/widgets/keyboard.py
@@ -146,6 +146,7 @@ class Key[CAT = str]:
     icon: ctk.CTkImage | None = None
     color: ctku.Color | None = None
     hover_color: ctku.Color | None = None
+    enabled: bool = True
 
 
 @dataclasses.dataclass
@@ -160,6 +161,7 @@ class KeyOverlay[CAT = str]:
     icon: ctk.CTkImage | None = None
     color: ctku.Color | None = None
     hover_color: ctku.Color | None = None
+    enabled: bool | None = None
 
 
 type OverlayLayerType[CAT = str] = dict[str | SpecialFunction | CAT, KeyOverlay[CAT]]
@@ -260,6 +262,8 @@ class Keyboard[CAT = str](ex.CTkFrameEx, AbstractRegistry):
                         image=key.icon,
                         fg_color=key.color,
                         hover_color=key.hover_color,
+                        state=ctku.flag_to_state(key.enabled),
+                        dark_when_disabled=True,
                         touchscreen_mode=self._touchscreen_mode,
                     )
                 button.configure(command=lambda k=key: self._handle_key(k))
@@ -293,12 +297,14 @@ class Keyboard[CAT = str](ex.CTkFrameEx, AbstractRegistry):
                 icon = overlay.icon if overlay.icon is not None else key.icon
                 color = overlay.color if overlay.color is not None else key.color
                 hover_color = overlay.hover_color if overlay.hover_color is not None else key.hover_color
+                enabled = overlay.enabled if overlay.enabled is not None else key.enabled
                 
                 button.configure(
                     text=self._select_button_text(value, text, icon),
                     image=icon,
                     fg_color=color if color is not None else ctk.ThemeManager.theme["CTkButton"]["fg_color"],
                     hover_color=hover_color if hover_color is not None else ctk.ThemeManager.theme["CTkButton"]["hover_color"],
+                    state=ctku.flag_to_state(enabled),
                 )
             else:
                 # otherwise just load the normal key values
@@ -307,6 +313,7 @@ class Keyboard[CAT = str](ex.CTkFrameEx, AbstractRegistry):
                     image=key.icon,
                     fg_color=key.color if key.color is not None else ctk.ThemeManager.theme["CTkButton"]["fg_color"],
                     hover_color=key.hover_color if key.hover_color is not None else ctk.ThemeManager.theme["CTkButton"]["hover_color"],
+                    state=ctku.flag_to_state(key.enabled),
                 )